+++ bin/build.php
@@ -9,6 +9,7 @@ function build_site(?Traversable<string> $filters = null): void {
 
   $steps = Vector {
     // No Dependencies
+    RoutingCodegenBuildStep::class,
     RawYAMLBuildStep::class,
     MergedYAMLBuildStep::class,
     FetchPHPDotNetIndexBuildStep::class,
