+++ composer.json
@@ -5,7 +5,9 @@
     "zendframework/zend-diactoros": "^1.1",
     "nikic/fast-route": "^1.0.1",
     "fredemmott/hhvm-autoload": "^1.0.0",
-    "fredemmott/type-assert": "^0.2.0"
+    "fredemmott/type-assert": "^0.2.0",
+    "fredemmott/hack-router-codegen": "^1.0.0rc2",
+    "fredemmott/hack-router": "^v1.0.0rc7"
   },
   "require-dev": {
     "fredemmott/definition-finder": "^1.1.0",
@@ -14,5 +16,6 @@
     "hack-psr/psr7-http-message-hhi": "^0.2.0",
     "phpunit/phpunit": "^5.1",
     "91carriage/phpunit-hhi": "^5.1"
-  }
+  },
+  "minimum-stability": "dev"
 }
