+++ src/site/HHVMDocumentationSite.php
@@ -1,5 +1,6 @@
 <?hh
 
+use FredEmmott\HackRouter as HackRouter;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Zend\Diactoros\Response\SapiEmitter;
@@ -13,12 +14,24 @@ final class HHVMDocumentationSite {
     (new SapiEmitter())->emit($response);
   }
 
+  private static function routeRequest(
+    ServerRequestInterface $request,
+  ): (classname<RoutableController>, ImmMap<string, string>) {
+    try {
+      return (new Router())->routePsr7Request($request);
+    } catch (HackRouter\NotFoundException $e) {
+      throw new HTTPNotFoundException('', 0, $e);
+    } catch (HackRouter\MethodNotAllowedException $e) {
+      throw new HTTPMethodNotAllowedException('', 0, $e);
+    }
+  }
+
   public static async function getResponseForRequest(
     ServerRequestInterface $request,
   ): Awaitable<ResponseInterface> {
     try {
       try {
-        list($controller, $vars) = (new Router())->routeRequest($request);
+        list($controller, $vars) = self::routeRequest($request);
       } catch (HTTPNotFoundException $e) {
         // Try to add trailing if we couldn't find a controller
         $orig_uri = $request->getUri();
@@ -26,8 +39,7 @@ final class HHVMDocumentationSite {
           ->withUri($orig_uri->withPath($orig_uri->getPath().'/'));
 
         try {
-          list($controller, $vars) = (new Router())->routeRequest(
-            $with_trailing_slash);
+          list($controller, $vars) = self::routeRequest($with_trailing_slash);
           // If we're here, it's routable with a trailing /
           return await (new RedirectException(
             $with_trailing_slash->getUri()->getPath()
