+++ src/site/RoutableController.php
@@ -1,0 +1,7 @@
+<?hh // strict
+
+use FredEmmott\HackRouter\IncludeInUriMap;
+
+interface RoutableController extends IncludeInUriMap {
+  require extends WebController;
+}
