+++ src/site/RoutableGetController.php
@@ -1,0 +1,8 @@
+<?hh // strict
+
+use FredEmmott\HackRouter\IncludeInUriMap;
+use FredEmmott\HackRouter\SupportsGetRequests;
+
+interface RoutableGetController extends
+RoutableController, SupportsGetRequests {
+}
