+++ src/site/Router.php
@@ -4,94 +4,12 @@ use FredEmmott\TypeAssert\TypeAssert;
 use HHVM\UserDocumentation\BuildPaths;
 use HHVM\UserDocumentation\LocalConfig;
 
-class Router {
-  private function getReadOnlyRoutes(
-  ): KeyedIterable<string, classname<WebController>> {
-    return ImmMap {
-      '/' => HomePageController::class,
-      '/search' => SearchController::class,
-      '/{product:(?:hack|hhvm)}/'
-        => GuidesListController::class,
-      '/{product:(?:hack|php)}/reference/'
-        => APIFullListController::class,
-      '/{product:(?:hack|php)}/reference/{type:(?:class|function|interface|trait)}/'
-        => APIListByTypeController::class,
-      '/{product:(?:hack|php)}/reference/{type:(?:class|function|interface|trait)}/{name}/'
-        => APIClassPageController::class,
-      '/{product:(?:hack)}/reference/{type:(?:class|interface|trait)}/{class}/{method}/'
-        => APIMethodPageController::class,
-      '/{product:(?:hack|hhvm)}/{guide}/'
-        => RedirectToGuideFirstPageController::class,
-      '/{product:(?:hack|hhvm)}/{guide}/{page}'
-        => GuidePageController::class,
-      '/manual/en/{legacy_id}.php'
-        => LegacyRedirectController::class,
-      '/robots.txt'
-        => RobotsTxtController::class,
-      '/s/{checksum}{file:/.+}'
-        => StaticResourcesController::class,
-      '/j/{keyword}'
-        => JumpController::class,
-    };
-  }
-
-  private function getWriteRoutes(
-  ): KeyedIterable<string, classname<WebController>> {
-    return ImmMap {
-      '/__survey/go'
-        => SurveyRedirectController::class,
-      '/__survey/nothanks'
-        => SurveyNoThanksController::class,
-    };
-  }
-
-  private function getDispatcher(): \FastRoute\Dispatcher {
-    return \FastRoute\cachedDispatcher(
-      function(\FastRoute\RouteCollector $r): void {
-        foreach ($this->getReadOnlyRoutes() as $route => $classname) {
-          $r->addRoute('GET', $route, $classname);
-        }
-
-        foreach ($this->getWriteRoutes() as $route => $classname) {
-          $r->addRoute('POST', $route, $classname);
-        }
-      },
-      /* HH_FIXME[4110] nikic/fastroute#83*/
-      shape(
-        'cacheFile' => BuildPaths::FASTROUTE_CACHE,
-        'cacheDisabled' => !LocalConfig::CACHE_ROUTES,
-      ),
-    );
-  }
-
-  public function routeRequest(
-    \Psr\Http\Message\ServerRequestInterface $request
-  ): (classname<WebController>, ImmMap<string, string>) {
-    $path = $request->getUri()->getPath();
-    $route = $this->getDispatcher()->dispatch(
-      $request->getMethod(),
-      $path,
-    );
-    switch ($route[0]) {
-      case \FastRoute\Dispatcher::NOT_FOUND:
-        throw new HTTPNotFoundException($path);
-      case \FastRoute\Dispatcher::METHOD_NOT_ALLOWED:
-        throw new HTTPMethodNotAllowedException($path);
-      case \FastRoute\Dispatcher::FOUND:
-        return tuple(
-          TypeAssert::isClassnameOf(
-            WebController::class,
-            $route[1],
-          ),
-          (new Map($route[2]))
-            ->map($encoded ==> urldecode($encoded))
-            ->toImmMap(),
-        );
+final class Router extends RouterCodegenBase {
+  <<__Override>>
+  protected function getCacheFilePath(): ?string {
+    if (LocalConfig::CACHE_ROUTES) {
+      return BuildPaths::FASTROUTE_CACHE;
     }
-
-    invariant_violation(
-      "Unknown fastroute result: %s",
-      var_export($route[0], true),
-    );
+    return null;
   }
 }
